+++ .codemod/classReplacements.js
@@ -1,4 +1,6 @@
-const classMap = {};
-
+const classMap = {
+  'xui-pageheading--breadcrumbs': 'xui-pageheading--breadcrumbtrail',
+  'xui-breadcrumbs': 'xui-breadcrumbtrail',
+};
 
 module.exports = classMap;